finish avatar change
