+++ Assets/Nova/Sources/Scripts/Graphics/CompositeSpriteController.cs
@@ -7,65 +7,27 @@ using UnityEngine;
 namespace Nova
 {
     [ExportCustomType]
-    public class CompositeSpriteController : OverlayTextureChangerBase, IRestorable, IOverlayRenderer
+    public abstract class CompositeSpriteController : FadeController, IRestorable
     {
         public const int mergerLayer = 16;
-        private const string overlayShader = "Nova/VFX/Overlay";
 
-        public GameObject overlay;
-        public GameObject overlayObject => overlay;
         public CompositeSpriteMerger mergerPrimary;
         public CompositeSpriteMerger mergerSub;
         public string imageFolder;
         public string luaGlobalName;
 
-        private List<string> curPose = new List<string>();
-        private static Mesh quad;
-        private MeshRenderer meshRenderer;
-        private MeshFilter meshFilter;
-        protected GameState gameState;
+        protected List<string> curPose = new List<string>();
         private DialogueState dialogueState;
-        private MyTarget myTarget;
-        private Material overlayMaterial;
+        protected GameState gameState;
 
+        public bool needRender => mergerPrimary.spriteCount > 0 || (isFading && mergerSub.spriteCount > 0);
         protected override string fadeShader => "Nova/VFX/Fade Global";
-        public RenderTexture renderTexture => myTarget == null ? null : myTarget.targetTexture;
+        public abstract bool renderToCamera { get; }
+        public abstract RenderTexture renderTexture { get; }
 
         protected override void Awake()
         {
-            if (quad == null)
-            {
-                quad = new Mesh();
-                quad.vertices = new[]
-                {
-                    new Vector3(-1, -1, 0),
-                    new Vector3( 1, -1, 0),
-                    new Vector3(-1,  1, 0),
-                    new Vector3( 1,  1, 0),
-                };
-                quad.uv = new[]
-                {
-                    new Vector2(0, 0),
-                    new Vector2(1, 0),
-                    new Vector2(0, 1),
-                    new Vector2(1, 1),
-                };
-                quad.triangles = new[]
-                {
-                    0, 2, 1,
-                    2, 3, 1
-                };
-                // some very large bound to disable culling
-                quad.bounds = new Bounds(Vector3.zero, 1e6f * Vector3.one);
-            }
-            meshFilter = overlay.Ensure<MeshFilter>();
-            meshFilter.mesh = quad;
-            meshRenderer = overlay.Ensure<MeshRenderer>();
             base.Awake();
-            overlayMaterial = materialPool.Get(overlayShader);
-            materialPool.defaultMaterial = null;
-            meshRenderer.material = overlayMaterial;
-
             var controller = Utils.FindNovaGameController();
             gameState = controller.GameState;
             dialogueState = controller.DialogueState;
@@ -75,9 +37,6 @@ namespace Nova
                 LuaRuntime.Instance.BindObject(luaGlobalName, this, "_G");
                 gameState.AddRestorable(this);
             }
-
-            myTarget = new MyTarget(this);
-            myTarget.Awake();
         }
 
         protected virtual void OnDestroy()
@@ -86,16 +45,14 @@ namespace Nova
             {
                 gameState.RemoveRestorable(this);
             }
-            myTarget.OnDestroy();
-        }
-
-        protected virtual void Update()
-        {
-            myTarget.Update();
         }
 
         public void SetPose(IEnumerable<string> pose, bool fade = true)
         {
+            if (curPose.SequenceEqual(pose))
+            {
+                return;
+            }
             fade = fade && !gameState.isRestoring && !dialogueState.isFastForward;
             if (fade)
             {
@@ -112,6 +69,11 @@ namespace Nova
             curPose.AddRange(pose);
         }
 
+        public void SetPose(string pose, bool fade = true)
+        {
+            SetPose(pose.Split(':'), fade);
+        }
+
         public void SetPose(LuaInterface.LuaTable pose, bool fade = true)
         {
             SetPose(pose.ToArray().Cast<string>(), fade);
@@ -120,6 +82,11 @@ namespace Nova
         public void ClearImage(bool fade = true)
         {
             SetPose(Enumerable.Empty<string>(), fade);
+        }
+
+        public static string PoseToString(IEnumerable<string> pose)
+        {
+            return string.Join(":", pose);
         }
 
         public virtual string restorableName => luaGlobalName;
@@ -161,34 +128,6 @@ namespace Nova
             color = data.color;
             gameObject.Ensure<RenderQueueOverrider>().renderQueue = data.renderQueue;
             SetPose(data.poseArray, false);
-        }
-
-        protected class MyTarget : RenderTarget
-        {
-            private new const string SUFFIX = "Composite" + RenderTarget.SUFFIX;
-            private CompositeSpriteController parent;
-            public override string textureName => parent == null ? oldConfig.name : parent.luaGlobalName + SUFFIX;
-            public override bool isFinal => false;
-            public override bool isActive =>
-                parent != null && (parent.mergerPrimary.spriteCount > 0 || (parent.isFading && parent.mergerSub.spriteCount > 0));
-
-            public override RenderTexture targetTexture
-            {
-                set
-                {
-                    base.targetTexture = value;
-                    if (parent != null)
-                    {
-                        parent.overlayMaterial.SetTexture("_MainTex", value);
-                        parent.overlay.SetActive(value != null);
-                    }
-                }
-            }
-
-            public MyTarget(CompositeSpriteController parent)
-            {
-                this.parent = parent;
-            }
         }
     }
 }
