+++ Assets/Nova/Sources/Scripts/Graphics/FadeController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 69039ebe4ed030543a8f9ed11e976031
+guid: 0bce201dfebbdc94d80b42fb16c85dce
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
