+++ Assets/Nova/Sources/Scripts/Graphics/OverlaySpriteController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: afd8c65760f4d5d4087905c4f69e2f4d
+guid: 309be5427461aaa459940a5bb17cbf7c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
