+++ Assets/Nova/Sources/Scripts/Graphics/TextureFadeController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fa5c3da214a441940a73990188ba4b1d
+guid: 9794603839ae61f4e925aedaea8ed85d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
