+++ ProjectSettings/TagManager.asset
@@ -8,7 +8,8 @@ TagManager:
   - UICamera
   - RenderManager
   - UIRoot
-  - SpriteMerger1
+  - Character
+  - Avatar
   layers:
   - Default
   - TransparentFX
